small refactors for fin provider
