+++ oracle/provider/fin.go
@@ -10,8 +10,6 @@ import (
 
 	"price-feeder/config"
 	"price-feeder/oracle/types"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 const (
@@ -109,15 +107,10 @@ func (p FinProvider) GetTickerPrices(pairs ...types.CurrencyPair) (map[string]Ti
 		if ok {
 			return nil, fmt.Errorf("FIN tickers response contained duplicate: %s", ticker.Symbol)
 		}
-		price := strToDec(ticker.Price)
-		if err != nil {
-			return nil, fmt.Errorf("FIN ticker price failed to parse for %s: %s", ticker.Symbol, ticker.Price)
+		tickerPrices[pair.String()] = TickerPrice{
+			Price: strToDec(ticker.Price), 
+			Volume: strToDec(ticker.Volume),
 		}
-		volume := strToDec(ticker.Volume)
-		if err != nil {
-			return nil, fmt.Errorf("FIN ticker volume failed to parse for %s: %s", ticker.Symbol, ticker.Volume)
-		}
-		tickerPrices[pair.String()] = TickerPrice{Price: price, Volume: volume}
 	}
 	for _, pair := range pairs {
 		_, ok := tickerPrices[pair.String()]
@@ -125,9 +118,6 @@ func (p FinProvider) GetTickerPrices(pairs ...types.CurrencyPair) (map[string]Ti
 			return nil, fmt.Errorf("FIN ticker price missing for pair: %s", pair.String())
 		}
 	}
-	fmt.Println("tickerPrices")
-	fmt.Println(tickerPrices)
-
 	return tickerPrices, nil
 }
 
@@ -142,10 +132,7 @@ func (p FinProvider) GetCandlePrices(pairs ...types.CurrencyPair) (map[string][]
 		if !ok {
 			return nil, fmt.Errorf("FIN contract address lookup failed for pair: %s", pair.String())
 		}
-		_, ok = candlePricesPairs[pair.Base]
-		if !ok {
-			candlePricesPairs[pair.String()] = []CandlePrice{}
-		}
+		candlePricesPairs[pair.String()] = []CandlePrice{}
 		windowEndTime := time.Now()
 		windowStartTime := windowEndTime.Add(-finCandleWindowSizeHours * time.Hour)
 		path := fmt.Sprintf("%s%s?contract=%s&precision=%d&from=%s&to=%s",
@@ -184,8 +171,6 @@ func (p FinProvider) GetCandlePrices(pairs ...types.CurrencyPair) (map[string][]
 		}
 		candlePricesPairs[pair.String()] = candlePrices
 	}
-	fmt.Println("candlePricesPairs")
-	fmt.Println(candlePricesPairs)
 	return candlePricesPairs, nil
 }
 
@@ -235,17 +220,6 @@ func (p FinProvider) getFinPairAddresses() (map[string]string, error) {
 // SubscribeCurrencyPairs performs a no-op since fin does not use websockets
 func (p FinProvider) SubscribeCurrencyPairs(pairs ...types.CurrencyPair) error {
 	return nil
-}
-
-func strToDec(str string) sdk.Dec {
-	if strings.Contains(str, ".") {
-		split := strings.Split(str, ".")
-		if len(split[1]) > 18 {
-			// sdk.MustNewDecFromStr will panic if decimal precision is greater than 18
-			str = split[0] + "." + split[1][0:18]
-		}
-	}
-	return sdk.MustNewDecFromStr(str)
 }
 
 func binToTimeStamp(bin string) (int64, error) {
